Let system choose appropriate integer types

+++ repa/pargrid.hpp
@@ -79,11 +79,11 @@ typedef int rank_type;
 
 /** Index of a local cell (0..n_local_cells-1).
  */
-using local_cell_index_type = util::StrongAlias<int, type_tags::LocalCellIndex>;
+using local_cell_index_type = util::StrongAlias<int_fast32_t, type_tags::LocalCellIndex>;
 
 /** Index of a ghost cell (0..n_ghost_cells-1).
  */
-using ghost_cell_index_type = util::StrongAlias<int, type_tags::GhostCellIndex>;
+using ghost_cell_index_type = util::StrongAlias<int_fast32_t, type_tags::GhostCellIndex>;
 
 /** cell_range.
  * Offers functions to conveniently iterate over a range of cells.
@@ -108,7 +108,7 @@ using local_or_ghost_cell_index_type
 /** Global cell index (unique across all processes).
  */
 using global_cell_index_type
-    = util::StrongAlias<int, type_tags::GlobalCellIndex>;
+    = util::StrongAlias<int_fast64_t, type_tags::GlobalCellIndex>;
 
 typedef std::function<std::vector<double>(void)> CellMetric;
 typedef std::function<double(local_cell_index_type,
